Escape single quotes while translating dropped Win32 paths
When file/folder is dropped to the terminal, its path is translated and
quoted with a pair of single quotes if necessary.

However, the terminal control does not escape single quotes (allowed in
the Win32 subsystem) that need escapes when translated.

On the translation styles other than "none" (note: all other translation
styles are currently intended for the POSIX shell), it causes incorrect
path to be pasted when the path contains one or more single quotes
(see Issue #18006 for an example).

With this commit, the terminal control escapes a single quote with a
valid escape sequence `'\''` (finish quote, print a single quote then
begin quote again) when the path translation is required.